feat: change function as util function
